add: trim() and make public to Board
